utils: fix time_now_ms() on 32bit platforms
This could overflow on platforms where time_t is defined to be 32bit.
Overflow would happen approximately after 24 days of uptime:

    0x80000000 / 1000ms / 60s / 60min / 24h ~= 24d

+++ common/time_extra.c
@@ -26,7 +26,7 @@ uint64_t time_now_ms(void)
     struct timespec now;
 
     clock_gettime(CLOCK_MONOTONIC, &now);
-    return now.tv_sec * 1000 + now.tv_nsec / 1000000;
+    return (uint64_t)now.tv_sec * 1000 + now.tv_nsec / 1000000;
 }
 
 time_t time_get_elapsed(clockid_t clockid, time_t start)
